requirements.txt
bcrypt
cerberus
python-arango
tornado
raven
tox
